refactor: move env logic to workload
